qcacld-3.0: Fix the reserved marking for TX flow entries
The flow_id stored in socket is 6-bit which indicates the hash_idx.
hash_idx=0 is reserved to indicate invalid flow id. This hash_idx=0
maps to 4 flow entries in the TX flow table, i.e. 0,64,12,192

Currently we are only marking flow_id=0 as reserved and the other
entries can get used for marking a flow. This causes issue, where
a particular flow getting these IDs will start using multiple flow
entries in the table. Also, when STC walks over the flow table to
do a bi-directional flow search, it is possible that STC picks one
of these invalid flow_ids as a match, since the flow_tuple will be
populated properly. This leads to invalid TX flow stats and
incorrect flow classification.

In order to fix the above mentioned issue, mark all the flow_ids
corresponding to hash_idx=0 as reserved and should not be used for
any TX flow mapping.

Change-Id: I2ec2396afc7eb11fbdd11928d2f3f8d357a0fbe6
CRs-Fixed: 4163749