qcacld-3.0: Don't send DISCONN event in case of connection failure
DISCONN connectivity logging event should be sent in case
of successful connection. But in some connecting fail scenario,
if the disconnect is received from userspace, it is trying to
send DISCONN logging event even though it is not successfully
connected. This issue occurs while iterating through candidates
for connection. Before the system can process all candidates,
a disconnect event with null bssid is triggered by userspace
due to the watchdog timer, leading to premature termination
of the connection attempt.
Drop the disconn event in this scenario by checking the BSSID.

CRs-Fixed: 4152755
Change-Id: Iba08bbb0117789c4fa6c4139fa7fa139dcfc03e2