qcacld-3.0: Separate API for Auto platform
pm_suspend_via_firmware is a deprecated API from kernel 6.12
We get a symbol missing error when we compile with Kernel_platform5.0
onwards.
Also pm_suspend_via_firmware is not required in Auto platform as it
is for deep sleep enablement.
Remove the call for pm_suspend_via_firmware in Auto platform.

Change-Id: I1ff3312ba1275dc584e6f1a3215baad307a3cdf4
CRs-Fixed: 4170991