qcacld-3.0: Reset flag on disconnect to prevent incorrect BW
When a disconnection occurs, the update_from_ap flag is not
reset, which can lead to incorrect bandwidth settings on
subsequent connections. For example, if the previous connection
was on a 6 GHz channel with 320 MHz bandwidth, this value
may be incorrectly reused due to the stale update_from_ap flag.

This change ensures that the update_from_ap flag is reset during
disconnection, preventing the reuse of outdated bandwidth
information and ensuring that the bandwidth is correctly updated
based on the current connection parameters.

Change-Id: I26513ee5afe4431d060145274164298a21e4fa8c
CRs-Fixed: 4158668