qcacld-3.0: Fix pkt IAT sum for TxRx stats
For first packet of the sample window, IAT should be considered 0 but
currently IAT is taken with reference to the previous packet which
may be belong to a prior window, thereby skewing the packet IAT sum
for the current window.

Update the logic to skip IAT accumulation for the first packet in each
sampling window, ensuring accurate inter-arrival time statistics.

Change-Id: I72e1d155442dd6e0def8c6c3429ad8df3c6270ee
CRs-Fixed: 4200188