qcacld-3.0: Reject the FASTREASSOC cmd based on the roam band config
Spplicant sends a FASTREASSOC command with a specific BSSID and
frequency to initiate roaming. However, the host driver does not
validate whether the frequency received via the REASSOC driver
command falls within the currently allowed roam bands configured
through QCA_ATTR_ROAM_CONTROL_BAND_MASK.

This can result in the roam invoke command being sent to the fw
even if the frequency provided in the command is not valid,
leading to a roam invoke failure.

Fix is to add a check in the host driver to validate that the
frequency received via the REASSOC command is within the band(s)
specified by the current roam band mask. If the frequency does
not match the configured roam bands, the FASTREASSOC command
should be rejected.

Change-Id: I5e43db2400f546c6feb31268ac6b749f761c5d35
CRs-Fixed: 4164703