qcacld-3.0: Use channel width stored in link_chan_info
For each link, the driver stores the channel width and
the corresponding scan PHYMODE in the struct wlan_channel *link_chan_info.

The issue arises because the host is recalculating the
channel width based on the scan PHYMODE. This approach
can lead to incorrect channel width values if all possible
scan PHYMODEs are not properly accounted for during manual
calculation.

To avoid such inconsistencies, the host should use the
channel width value already stored in the link_chan_info
table, rather than recalculating it manually.

Change-Id: I9c58b7c0c6b765721c75ca9a53756f86bd6c0678
CRs-Fixed: 4167649