qcacld-3.0: Log stats only nl_app is active state
Currently, station statistics for the unicast command are being
logged regardless of whether a netlink application is registered.
This can lead to confusion about whether the application is
still active.

To address this, statistics should only be logged when the
application is confirmed to be alive and actively listening.

Change-Id: I29c88b1d605af6c5c05a2329fedb391661c1cec2
CRs-Fixed: 4152484