qcacld-3.0: Update policy manager PCL tables for PM_LL_LT_SAP
When XPAN comes up first, followed by the MLO 2-link STA, and
then NAN starts, the current behavior allows NAN to come up.
However, the expected behavior is that LT_SAP and NAN
concurrency should not be allowed.

This change updates the PCL table logic to ensure that if
LT_SAP is already active, NAN is not allowed to start, and
vice versa.

Change-Id: I9aae87058895d291adbbf83bf4151cbc9864b641
CRs-Fixed: 4182080