qcacld-3.0: Retain the roam context until osif update
Scenario:
1) DUT roams using EAPOL Suite-B AKM with 2 links.
2) For this AKM, EAPOL handling is offloaded to the
supplicant.
3) Therefore during roam sync, host driver brings up
the assoc vdev and updates the link id of all the
links in sta_ctx.
4) After EAPOL completion, driver tries to bringup
the partner vdev. During the cleanup of the old
partner vdev, the link-id from sta_ctx is cleared out
for vdev-1.
5) As part of the successful bring up of the new partner
vdev, the sta_ctx should be updated with the link ID.

To fix this, prevent clearing the roaming context before
osif update, and add a callback to clear the roaming
context after osif update.

Change-Id: I1e9278d4139c737f39c68c306610933418a8ed7e
CRs-Fixed: 4184876