qcacld-3.0: Exempt the SAP restart if NAN is present
Currently, the driver tries to restart SAP to the user-configured
frequency after a STA disconnect, even when SAP is not standalone
and runs concurrently with NAN. This causes NDP to fail if both
SAP and NDP use the 5 GHz band.

To fix this, the driver skips the SAP restart when NAN is active.

Change-Id: Id3ffbb6aa5ada700f44aa5de025e95d354a0a27b
CRs-Fixed: 4179325