qcacld-3.0: Improve the Tx-Rx flow eviction logic
With STC enabled, the flow which are selected to sample
are configured with an inactivity timeout of 10 seconds.
For bursty traffic, especially flows which are not balanced
across both directions, i.e. having sparingly less traffic
in one direction, it is very common/easy to hit the 10 seconds
inactivity timeout. This causes the flows to be retired midway
leading to incorrect stats collected by STC during sampling.

In order to fix this, increase the inactivity timeout for
Tx-Rx flows to 30 seconds, if selected to sample. This will
allow sufficient time for the STC sampling to complete.

Also, improve the flow eviction logic for Tx-Rx flows taking
STC into account.

Change-Id: I5087ce0e7adc00a0e75e8abb487edb6a3ae457e9
CRs-Fixed: 4188280