qcacld-3.0: Allow CSA when GO NOA duration is less than 500 ms
When STA scan is active, firmware enables an one-shot NOA on the P2P GO
interface. The current driver rejects CSA requests if any one-shot NOA
is active. To reduce the impact of this behavior, optimize the logic
to allow CSA when the NOA duration is less than 500 ms.

Change-Id: I722c76c9a69c9f2ea7b5b96349ef128bfc17cbbb
CRs-Fixed: 4214292