qcacld-3.0: Handle peer_id change in rx_flow after roaming
STC tracks the Tx-Rx flows, using RX flow as anchor and
uses the peer_id from RX flow table to map the sampled
flow to a particular peer.

After roaming is done, the traffic
flows corresponding to the deleted peer are removed from the
STC tables. Once these flow resumes after roaming, these are
sampled again, but the peer_id in the RX flow table is not updated.
This misleads STC to tag these flow against an invalid peer.

In order to fix the above mentioned issue, mark the peer_id
as INVALID, after the peer gets deleted, and update the new peer_id
when the flow resumes, post roaming.

Change-Id: I9cbb24924b8765fb4e324a687038c25d9093db6a
CRs-Fixed: 4186944