qcacld-3.0: Fix crash due to missing LL LT SAP entry during FW bearer switch
Previously, there was no handling for firmware bearer switch
requests when the LL LT SAP vdev was UP but not added to the
policy manager table. This led to crashes in the firmware.

This update adds proper handling for such scenarios, ensuring
stability even when the vdev is active but not registered in
the policy manager.

Change-Id: I5d3c6f72da1a4d0d7c586cf1806aeb392968a930
CRs-Fixed: 4187454