qcacld-3.0: Stop SAP if STA is connected with WAPI AP
In existing scenario, if SAP tries to come where STA is
already connected with WAPI AP, then driver rejects the
SAP to start due to WAPI security restrictions.

But in case of SAP + STA where SAP starts first, then
STA attempts to connect to a WAPI AP. Initially, the
WAPI security flag is not set, so driver allows the
STA connection. Once the STA connection completes, both
SAP & STA interfaces are in MCC. Driver calls
policy_mgr_is_concurrency_allowed() to move SAP to
SCC channel. But it fails the concurrency check and both
SAP & STA remains in MCC.

As part of fix, stop SAP if STA is connected with WAPI AP.

Change-Id: Ie80f78d0b84f3e7a7dc82221d00598f707a628ae
CRs-Fixed: 4183607