qcacld-3.0: Add support to handle SUBCMD_FEATURE_CONFIG
This change is to handle SUBCMD_FEATURE_CONFIG vendor command.
With this change vendor command will be parsed and
below validations will be done.

I.) CONFIG ACTION and CONFIG DATA OP are manadatory parameters,
if these are not sent by userspace then rest params will
not be parsed and failure will be return to userspace.

II.) IF DATA or MAC address is present but mask is not
present then parsing will not be done and failure will
be return to userspace.

III.) IF both CONFIG ACTION and CONFIG DATA OP are present without any
other param then don't disable DSMPS.

IV.) IF Config Data list is provided for more than 100 which
is MAX LIMIT then process DSMPS disable without saving the
pending OUI and return success to userspace.

V.) Don't process the request for 6 GHz band.

VI.) Don't process the request for EHT or HE Cap

Change-Id: I610e01cd8b99a358ef483c7caee3a098b4387f17
CRs-Fixed: 4192874