qcacld-3.0: Fix offchannel BW downgrade in TDLS
6 GHz offchanel TDLS cannot be established on a
bandwidth more than 80 MHz. This is due to removal
of the 160 MHz cap from the ini config. The 160 MHz
cap is removed based on the he cap, however the he
cap is not initialized with the target cap during
psoc open.

To fix this, intersect the target max bandwidth cap
with the ini config during off channel TDLS bw
computation instead.

Change-Id: I6ee0d893bc2af56c765c287929e125fdbb9761ef
CRs-Fixed: 4190409