qcacld-3.0: Add feature to restart SAP if sub channel is unsafe
Currently Coex safe channel check in wlansap_get_chan_band_restrict()
only apply to SAP’s primary channel, there is a new requirement to
check all sub channels in bandwidth whether they are safe to operate,
because on some platform it is easy to have short unsafe frequency
range.

Add a new feature to check all sub channels in bandwidth, call
wlansap_get_safe_channel_from_pcl_and_acs_range() to get target
channel frequency and bandwidth and switch to it if there is any
sub channel in bandwidth unsafe. Because
policy_mgr_is_sap_freq_allowed() have different requirements to
judge channel safety, it is not applicable to add bandwidth
to it and other platform may not need this feature, so add a
new feature flag for it, which only enable on sdxkova.

Change-Id: I5eb30491dcaef6598596abd863c15f779a369833
CRs-Fixed: 4173736