qcacld-3.0: Skip LPC suspend if LPC is not running
FW acquires a wakelock when Local Packet Capture is
enabled and releases it when LPC is disabled. If the
LPC suspend command is issued while LPC is not running,
The FW attempts to release a wakelock that was never
acquired, upon the reception of the
HTT_H2T_MSG_TYPE_TX_MONITOR_CFG message.This leads
to a firmware crash.

This change adds a check which skip suspend or disable
operations if LPC is already disabled.

Change-Id: I03fa120be845cc6b577a072a507ad441ba524b6b
CRs-Fixed: 4201677