qcacld-3.0: Fix iova ref count issue for page pool buffers
IOVA reference counters used for mapping page pool buffers into IPA
IOMMU domain requires the buffers to be perfectly aligned either
to the 2KB/4KB boundary, if the buffers are not aligned to this
requirement, IOVA reference counting logic will not work and this
will result unwanted system behavior.

In the chipsets that do not support QWORD based buffer optimizations,
size of the RX buffers will be 2.5KB which is not aligned to 2KB/4KB
need by IOVA refcount logic, therefore IPA IOMMU mapping for such
buffers will fail. Align the buffers to 4KB boundary to avoid the
failures.

CRs-Fixed: 4210452
Change-Id: I5cb3f614bf4cd67a1df0df77c7afc4ec5467e35f