qcacld-3.0: Fix excessive printk during authentication
The host driver currently uses printk for authentication frames.
If a peer continuously sends authentication frames, it causes
delays in the scheduler thread. To mitigate this, the driver
now uses rate limiting to prevent excessive logging. Additionally,
the retry bit is not set when the peer sends packets with the same
sequence number. The driver will check if the sequence has already
been processed and drop duplicate packets.

Change-Id: Ied4d91b9903df83c787ddf74da6853160676faae
CRs-Fixed: 4205615