qcacld-3.0: Move SAP mandatory channel to per vdev
Currently, the host driver uses a mandatory channel selection
based on the policy manager context, which is global. However,
for dual SAP configurations, this can be overridden if the 6 GHz
SAP comes up first and the 2 GHz SAP comes up later, causing the
6 GHz channel to be overridden by the 2 GHz channel.

This change ensures that the mandatory channel is managed per vdev,
preventing it from being overridden and maintaining it per vdev.

Change-Id: I71a4c33cae83c465d14d0c5d5e958f4840c4d4f7
CRs-Fixed: 4206457