qcacld-3.0: Fix key cleanup during stop BSS
This change addresses an issue where the host receives
multiple connect and disconnect events, but the
associated keys are not being deleted during the stop
BSS process. To resolve this, a conditional check is
added to ensure that key cleanup and beacon memory
release only occur when the driver is not recovering
or unloading.

CRs-Fixed: 4211784
Change-Id: I7fb658b3fcce5e8f88493a794bdbc0e7d915f5af