qcacld-3.0: Update the LL LT SAP freq after start bss is active
In case of STA and XPAN starting together, there is a chance that
STA connects before the LL LT SAP on the same freq as selected by
LL LT SAP.

This lead to SCC in LL LT SAP and later assert in firmware.

To fix, reevaluate the LL LT SAP freq again after start bss
become active in serilization.

Change-Id: I1c2c375f52030dc8833e6c77693fd1ef8c0e305f
CRs-Fixed: 4213404