qcacld-3.0: Fix get RX page pool failures from prealloc pool
Currently, number of RX page pools and the size of each pool are
decided based on WLAN_DP_DYNAMIC_RESOURCE_MGMT compilation flag
irrespective of whether the dynamic RX buffer allocation feature
is enabled from the INI or not.

When the dynamic RX buffer allocation feature is disabled from
the INI, RX page pool get operation from the prealloc pool fails
because of mismatch in size of the pool allocated during prealloc
and the size of pool requested during get operation.

Change the logic to allocate RX page pools based on the INI value
to fix this issue.

In addition, code refactoring is also done to avoid duplication.

CRs-Fixed: 4221342
Change-Id: I9bcf23ac4e0434167c43f83624abfae4df569938