qcacld-3.0: Add SAP/GO SCC check to prevent MCC scenarios
Currently, the host driver does not perform an SCC
check for SAP/GO when a CSA is triggered by the user.

This can lead to an unintended MCC scenario when two SAP
interfaces are active.

This change introduces an SCC check for SAP/GO during user-triggered CSA
to ensure proper channel coordination and avoid MCC conditions.

CRs-Fixed: 4220731
Change-Id: I8be2dcd330144c1cae0ece0330638641117b0de9