qcacld-3.0: Enable idle roam only when both PS and suspend are set
Idle roam enable is sent to firmware
based on the order of SETSUSPENDMODE and
NL80211_CMD_SET_POWER_SAVE commands.
Since the order is not guaranteed, this causes
a race condition where the driver does not update
idle roam status correctly to firmware.

This change introduces a 2-bit QDF bitmap
in the driver to track power save and suspend mode
states independently. The driver sends idle roam
enable to firmware only when both bits are set,
and sends disable only when both are cleared.

This ensures order-independent handling of
suspend and power save commands from userspace.

Change-Id: Id1c0db8c96d5890dd84b8f1ac179976b1ee97bb3
CRs-Fixed: 4188272