qcacld-3.0: Fix STA CSA on invalid 6 GHz frequency
If the AP sends a CSA on a frequency that falls outside the valid
range defined by the current country’s regulatory rules, the STA may
still allow the connection. This results in a regulatory violation.

For example, the STA initially connected on frequency 5975 using the
VLP power mode. It then received a CSA to switch to frequency 7115,
which is not valid for VLP power mode as per the current country
regulations. Currerntly still allowing the STA to new frequency.

To resolve this, ensure that the correct power mode is passed in the
CSA. This allows the system to determine the appropriate channel
state based on the current power mode and comply with regulatory
requirements.

Change-Id: I336279fcac8752022c0a523e342e59bb22ca40af
CRs-Fixed: 4223863