qcacld-3.0: Align SAP bandwidth with concurrent STA connection
Previously, when SAP was started on a 320 MHz channel and
a concurrent STA connection was present with lower bandwidth,
the host driver would downgrade SAP to 20 MHz.

This change improves bandwidth handling by adjusting SAP to
match the bandwidth of the concurrent STA connection,
rather than defaulting to 20 MHz.

CRs-Fixed: 4415574
Change-Id: Ia7efbc315afb2d657c2eba51ed8f659a42a734a8