qcacld-3.0: Send correct OWE link ID to kernel in non-MLO
In non 11be targets, link ID parameter of OWE info element is not
assigned -1 to indicate a non-MLO case. Instead, it defaults to 0
which is a valid MLO link ID. This occurs because the function
that fills OWE info is a placeholder with no logic in the non-MLO
flow.

To address this, modify the placeholder function to initialize
link ID with -1 value in non-MLO case.

CRs-Fixed: 4224782
Change-Id: Iee6e86d39cb3bc414674b5455a617e793558091f
(cherry picked from commit 95887378d22bb7d793722f6dbf847720fc11d30a)