qcacld-3.0: Fix race condition for set vdev IEs
The scheduler is dequeued to populate the vdev IEs, but at
this point, the vdev has already been destroyed. Meanwhile still
iterating over each dot11mode. During this process, vdev is
created with existing vdev_id, but the NSS has not yet been
initialized. As a result, NSS remains 0 when attempting to
retrieve the data rate, leading to an array access with index -1,
which causes an OOB access.

To resolve this race condition, add a check to ensure the vdev is
valid before populating the IEs. This prevents accessing invalid
memory and avoids the OOB issue.

Change-Id: Ie782d6faa44b7a603bdcbbf3721458149ef9b349
CRs-Fixed: 4247030