qcacld-3.0: Add WFA test command to configure Extended MLD cap to FW
STA Host has different handling of Extended MLD capability support in
association request for certification cases and on-field behavior. For
Wi-Fi 7 cert, the Ext MLD caps are excluded by default for Testbed STA,
and enabled by default for STAUT test cases. For cases that need
Ext MLD cap to be set (e.g., Load balancing/Link reconfiguration),
Host has a vendor command to set Ext MLD caps in association req. For
on-field cases, STA has to check if AP advertises Ext MLD caps in
beacons or if it's a 3-link association to set the self caps to avoid
IOT issues.
Firmware does not know if it's for cert case or on-field to set the
cap appropriately in reassociation request.
Hence, Add WFA test command support to configure Extended MLD
capability support to FW as the support varies with STA role and
vendor command.

CRs-Fixed: 4284887
Change-Id: If4b6e9816f5fde055f15529151df47bcfafa2a0d