qcacld-3.0: Fix the bearer switch SM callback corruption
In the event of a bearer switch SM loop, it is essential to handle
callbacks carefully. If callbacks are looped and set to NULL, a
new request might be added to the end of the callback array during
the callback process. This could result in the driver calling the
newly added callback and setting it to NULL before the new callback
has finished processing for that request.

To prevent this race condition, create a local copy of the callbacks,
reset the original callbacks, and use the local copy to execute the
callbacks. This approach ensures that new callbacks are not
prematurely set to NULL.

Change-Id: I554147d71038ef7ba235c08558fdfb947457f867
CRs-Fixed: 4248538