qcacld-3.0: Fix incorrect SAP puncture bitmap after restart
When DFS occurs and sub-channels in the regulatory domain are disabled,
SAP is stopped and restarted. However, the puncture bitmap in the PE
session was not updated to reflect the disabled channels, resulting in an
incorrect puncture bitmap in the EHT Operation element of the beacon
template.

This fix ensures that, upon SAP start, the puncture bitmap is updated based
on the disabled channels in the regulatory domain and refilled into the PE
session. The puncture bitmap is then updated in the EHT Operation element
and finally populated into the beacon template.

Change-Id: Ie8bf92e7af0bdde449d34e6b1f03de394965b286
CRs-Fixed: 4258380