qcacld-3.0: Prevent crash when STC is enabled and FISA is disabled
When FISA is disabled through configuration and STC is enabled,
the system attempts to use RX flow table resources that are not
initialized, which leads to a crash during STC operations.

With FISA disabled via INI and STC enabled, initialize the RX flow
table and allow flow table entry operations which are required by
STC to function properly. FISA aggregation will not be done
in such a scenario.

Change-Id: I06e70f05bc1c7b2fc1a569344ab037bf9706a5cc
CRs-Fixed: 4257527