qcacld-3.0: Fix the wpa3 roam disconnect issue when set sae_pwe 1
This disconnect is caused by the peer AP send the deauth after
received our STA's EAPOL-2. The reason is that our STA's EAPOL-2
missed the RSNXE IE in the key data when we set the sae_pwe=1.
As this is LFR2.0, the key data in EAPOL-2 are filled in the
wpa_supplicant, but from the wpa_supplicant, the RSNX IE fetched
from the assoc req ie in the asssoc event which sent from the
wlan driver.
So adding the assoc req ies which contains the RSNXE IE in the
asssoc event for the SAE H2E(sae_pwe=1) case in wlan driver,
and then wpa3 roaming successfully in this case.

Change-Id: I833a67344a65f2507b31a37a6e41ac5f90e957ef
CRs-Fixed: 4232383