qcacld-3.0: Add 6 GHz RNR IE in lim_send_probe_rsp_mgmt_frame
In some cases, it is necessary to check the 6 GHz RNR IE in probe
responses sent by SAP. Currently, both beacon and probe response
templates are included. However, the probe response may not always
be offloaded and sent by the target—it might instead be handled
by the host, such as when a probe request contains a WPS IE.

Therefore, the host must also populate the 6 GHz RNR IE in the
probe response, just as it does for the beacon template.

Change-Id: Id6b0dcd102323d42d452236841cc216f20a57d1d
CRs-Fixed: 4264424