qcacld-3.0: Improve connectivity logging for WEP
The authentication algorithm for the connectivity
logging is determined by referring to the first
byte after mac header offset. However, for the WEP
frames which are encrypted, the WEP control header
would be present after mac header offset. This leads
to populating junk values from the first byte of the
WEP header as the authentication algorithm.

To fix this, if the authentication frame is encrypted,
mark the authentication algorithm as WEP.

Change-Id: I376296fb5f1ded7550f89e82b96bf1ed183e5450
CRs-Fixed: 4267949