qcacld-3.0: Enable PHY Mode Change via Self-Roaming
Allows PHY mode to be changed while the STA is connected using a vendor
command. The change takes effect by invoking self-roaming with the
following command:
wpa_cli DRIVER FASTREASSOC BSSID FREQ

After connecting using MLO (6+2), if the PHY mode is downgraded to Wi-Fi 5
and self-roaming is triggered to a 2.4 GHz legacy link, the following issue
occurs:
Upon receiving the authentication response from the AP, the vdev0 MAC
address remains the link address, while the authentication destination
address is the MLD address. As a result, the system cannot locate the
correct vdev0 using the MLD address. It then attempts to find the vdev
by BSSID, but the 2.4 GHz BSSID is still associated with the vdev1 PE
session. This leads to retrieving the wrong PE session, which is in an
incorrect state. Consequently, the authentication response is dropped,
authentication fails, roaming fails, and a disconnect occurs.

To resolve this, when receiving an authentication response during roaming,
always use the associated vdev to handle authentication frames.

Change-Id: Ie2c232dc08e75cdbc01c87f0216101f311dd3dea
CRs-Fixed: 4274407