qcacld-3.0: Allow fixed channel SAP to operate on unsafe channels
Fixed channel SAP is allowed to operate even when a channel is
marked as unsafe due to coex operations.
As part of SAP+SAP+ML-STA support, SAP restart is considered to
be in SCC with the ML-STA links. But above constraint is not
considered. So, this results into moving the SAP out of unsafe
channel always when there is no concurrent STA.
Skip fixed channel SAP CSA even if the channel is marked as
unsafe.

Change-Id: I5725304ac93095c839174d8c9a0682fd0b6cd6e4
CRs-Fixed: 4254439