qcacld-3.0: Update link bssid before logging rssi and snr info
In the current code flow, the API responsible for logging RSSI
and SNR information is invoked before the link information is
saved. As a result, the logger receives an invalid BSSID as
input, leading to unexpectedly high RSSI values being logged
from this incorrect BSSID.

To address this, call the logger API after properly saving the
link information.

CRs-Fixed: 4259472
Change-Id: I1676b5b0230b3a64e41e6155774d41f133a3c5eb