qcacld-3.0: Check channel width before packaging EHT CAP for link
When the MLO SAP is started with 6 GHz band 80 MHz bandwidth as the base
link and 5 GHz band 160 MHz bandwidth as second link, the EHT CAP of the
base link and the second link will be exactly the same. This cause the
STA failed to parse second link's EHT CAP when connecting and parsing
association response frame, because its bandwidth is 160 MHz and base
link's EHT CAP is packed based on 80 MHz.

Change-Id: I37041169a0edaa376774f062c9f7c67e9f1ce354
CRs-Fixed: 4284049