qcacld-3.0: Forcefully bringup SAP on 20 MHz for country: ID
Forcefully bringup SAP on 20 MHz for country: ID
and gForceSAP20Mhz is enabled
Below are the cases:
1. Start SAP on 20 MHz:
SAP is started directly on 20 MHz due to regulatory
constraints for country ID.
2. During country code changes:
When the country code is dynamically updated to ID,
SAP is reconfigured to operate on 20 MHz.
3. During CSA trigger:
If a CSA is triggered while operating under country ID,
the driver ensures SAP remains on 20 MHz.

Change-Id: Ibedc69b0a3cb2bd4c6c1c92b8ed2a6368cda1070
CRs-Fixed: 4282121