qcacld-3.0: Allow 5 GHz scan when DFS SAP SCC with STA is enabled
When SAP (GO) operates in SCC mode with STA on a DFS channel and this
configuration is permitted via INI settings, the 5 GHz scan channel list
should not be filtered. In this scenario, the DFS master role is
delegated to the STA interface.
This change adds a SCC check for DBS-capable hardware and dfs master of
SAP not enabled ensures policy_mgr_scan_trim_5g_chnls_for_dfs_ap()
returns false, allowing 5 GHz scans to proceed.

Change-Id: Ie113f71902d38e185c3e7dd18d505277092aa94b
CRs-Fixed: 4286940