qcacld-3.0: Fix incorrect MCS rate count for 1x1 STBC-disabled HT 1x1 AP
When connecting to a 1x1 STBC-disabled HT 1x1 AP, the valid peer MCS rate
count should be 8 (MCS0–7), but the peer association code incorrectly
sends 9 rates to the firmware, treating all 128 bits of the HT Supported
MCS Set field as Rx MCS Bitmask. This leads to the firmware overriding the
Tx STBC bit in the HT capabilities for out of buffer issue happens when
handling the 9th invalid peer rate, causing STBC to be enabled in TX
packets. As a result, the AP cannot receive or acknowledge these packets.

This fix ensures that only the Rx MCS Bitmask portion (the first 77 bits
of the HT Supported MCS Set field) is checked when determining the MCS
rate count.

Change-Id: Ib43d5cd49be8e004b03bf07d617c442a9f66255b
CRs-Fixed: 4282364