qcacld-3.0: Populate SSID in SAP Mode for LL STATS
The SSID field in the wifi_interface_info is not
populated for SAP interfaces; only the BSSID is
being filled. This results in missing SSID information
when querying LL stats in SAP Mode.

Added code to populate the SSID field for SAP mode
by copying the SSID from the SAP configuration.
This ensures that both SSID and BSSID are available
in the wifi_interface_info so that SSID will also be
populated for LL stats command.

Change-Id: I15873c0753ce1aaa14a7df3529e01260a3c806e0
CRs-Fixed: 4267151