qcacld-3.0: Fix roam failure due to duplicate BSSID
Scenario:
1) FW does 2 link ML roaming with partner bringup offloaded
to the host.
2) The SSID of the roaming is determined from the scan entry
retrieved based on the bssid of the assoc link.
3) If there is a duplicate scan entry with same BSSID, then
the ssid retrieved would be different than the actual SSID.
4) The SSID from scan entry is saved into vdev, and used for
partner bringup which fails to find the candidate.

To fix this, always use bssid and ssid from the reassoc request
to retrieve the actual scan entry of the candidate.

Change-Id: I79c2e482ee2645e332ed9f62bfe480f37a7246a2
CRs-Fixed: 4295187