qcacld-3.0: Correct Reporting of max Bandwidth for Standby Link
In scenarios where the 5 GHz band is used as a standby link and the
user configures dynamic bandwidth to 320 MHz, the driver correctly
rejects the command due to hardware limitations. However, the
get_maximum_bandwidth command still returns the user-configured
value, leading to a misleading result and poor user experience.

Since the maximum supported bandwidth for 5 GHz is 160 MHz, the
driver should detect and report the invalid channel width instead
of echoing the unsupported configuration.

Change-Id: Ibd1713f733f2402ee14d4e47c014a6ed48df4a59
CRs-Fixed: 4292568